+++ changelog.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog][keep a changelog] and this project adh
 
 ## [Unreleased]
 
+### Added
+
+- Add user dashboard component and button to library exports.
+
 ### Fixed
 
 - Coupon field styling issues in some production environments.
+- Translation improvements.
 
 ## [Released]